feat(exceptions): custom exception for unauthorized database operations

+++ src/main/java/com/sgallalucas/gym/services/ProfessorService.java
@@ -1,12 +1,16 @@
 package com.sgallalucas.gym.services;
 
+import com.sgallalucas.gym.exceptions.NotAllowedOperationException;
 import com.sgallalucas.gym.model.Professor;
+import com.sgallalucas.gym.model.Student;
+import com.sgallalucas.gym.model.Workout;
 import com.sgallalucas.gym.repositories.ProfessorRepository;
 import com.sgallalucas.gym.validators.ProfessorValidator;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -38,6 +42,15 @@ public class ProfessorService {
 
     public void delete(Professor professor) {
         Professor found = getProfessor(professor.getId());
+        if (hasStudentOrWorkout(found)) {
+            throw new NotAllowedOperationException("It's not allowed delete a professor that has students or workouts");
+        }
         professorRepository.delete(found);
+    }
+
+    public boolean hasStudentOrWorkout(Professor professor) {
+        List<Student> studentList = professor.getStudents();
+        List<Workout> workoutList = professor.getWorkouts();
+        return (studentList.isEmpty() || workoutList.isEmpty()) ? false : true;
     }
 }
